pyproject.toml
[project]
name = "pr-splitter"
version = "0.1.0"
description = "Split large PRs into smaller, file-based PRs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "gitpython>=3.1",
    "pydantic>=2.0",
    "pathspec>=0.12",
]

[project.scripts]
pr-splitter = "pr_splitter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_splitter"]

[tool.mypy]
strict = true
python_version = "3.13"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "mypy>=1.10",
]
